# swagger/doc.yml
swagger: "2.0"
info:
  description: "A GinBro RESTful APIs"
  version: "1.0.0"
  title: "GinBro RESTful APIs Application"
host: "127.0.0.1:5555"
basePath: "/api/v1"
schemes:
- "http"
paths:
...
  /basepayroll:
    get:
      tags:
      - "basepayroll"
      summary: "get all basepayroll by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/BasepayrollPagination"
    post:
      tags:
      - "basepayroll"
      summary: "create basepayroll"
      description: "create basepayroll"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create basepayroll"
        required: true
        schema:
          $ref: "#/definitions/Basepayroll"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "basepayroll"
      summary: "update basepayroll"
      description: "update basepayroll"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create basepayroll"
        required: true
        schema:
          $ref: "#/definitions/Basepayroll"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /basepayroll/{Id}:
    get:
      tags:
      - "basepayroll"
      summary: "get a basepayroll by ID"
      description: "get a basepayroll by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of basepayroll to return"
        required: true
        type:  "string" 
        format:  "string"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Basepayroll"
    delete:
      tags:
      - "basepayroll"
      summary: "Destroy a basepayroll by ID"
      description: "delete a basepayroll by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of basepayroll to delete"
        required: true
        type:  "string" 
        format:  "string"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
...
  /personpayroll:
    get:
      tags:
      - "personpayroll"
      summary: "get all personpayroll by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/PersonpayrollPagination"
    post:
      tags:
      - "personpayroll"
      summary: "create personpayroll"
      description: "create personpayroll"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create personpayroll"
        required: true
        schema:
          $ref: "#/definitions/Personpayroll"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "personpayroll"
      summary: "update personpayroll"
      description: "update personpayroll"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create personpayroll"
        required: true
        schema:
          $ref: "#/definitions/Personpayroll"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /personpayroll/{Id}:
    get:
      tags:
      - "personpayroll"
      summary: "get a personpayroll by ID"
      description: "get a personpayroll by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of personpayroll to return"
        required: true
        type:  "string" 
        format:  "string"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Personpayroll"
    delete:
      tags:
      - "personpayroll"
      summary: "Destroy a personpayroll by ID"
      description: "delete a personpayroll by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of personpayroll to delete"
        required: true
        type:  "string" 
        format:  "string"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /tenant:
    get:
      tags:
      - "tenant"
      summary: "get all tenant by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/TenantPagination"
    post:
      tags:
      - "tenant"
      summary: "create tenant"
      description: "create tenant"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create tenant"
        required: true
        schema:
          $ref: "#/definitions/Tenant"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "tenant"
      summary: "update tenant"
      description: "update tenant"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create tenant"
        required: true
        schema:
          $ref: "#/definitions/Tenant"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /tenant/{Id}:
    get:
      tags:
      - "tenant"
      summary: "get a tenant by ID"
      description: "get a tenant by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of tenant to return"
        required: true
        type:  "string" 
        format:  "string"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Tenant"
    delete:
      tags:
      - "tenant"
      summary: "Destroy a tenant by ID"
      description: "delete a tenant by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of tenant to delete"
        required: true
        type:  "string" 
        format:  "string"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /tenantadjustitem:
    get:
      tags:
      - "tenantadjustitem"
      summary: "get all tenantadjustitem by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/TenantadjustitemPagination"
    post:
      tags:
      - "tenantadjustitem"
      summary: "create tenantadjustitem"
      description: "create tenantadjustitem"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create tenantadjustitem"
        required: true
        schema:
          $ref: "#/definitions/Tenantadjustitem"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "tenantadjustitem"
      summary: "update tenantadjustitem"
      description: "update tenantadjustitem"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create tenantadjustitem"
        required: true
        schema:
          $ref: "#/definitions/Tenantadjustitem"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /tenantadjustitem/{Id}:
    get:
      tags:
      - "tenantadjustitem"
      summary: "get a tenantadjustitem by ID"
      description: "get a tenantadjustitem by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of tenantadjustitem to return"
        required: true
        type:  "string" 
        format:  "string"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Tenantadjustitem"
    delete:
      tags:
      - "tenantadjustitem"
      summary: "Destroy a tenantadjustitem by ID"
      description: "delete a tenantadjustitem by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of tenantadjustitem to delete"
        required: true
        type:  "string" 
        format:  "string"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
...
definitions:
...
  Basepayroll:
    type: "object"
    properties:
    
      pid:
        type: "string"
        description: "编号"
        format: "string"
      
      standard_name:
        type: "string"
        description: "标准名称"
        format: "string"
      
      field_name:
        type: "string"
        description: "字段名称"
        format: "string"
      
      formula:
        type: "string"
        description: "日薪公式"
        format: "string"
      
      formula_mem:
        type: "string"
        description: "公式备注"
        format: "string"
      
      is_must_used:
        type: "integer"
        description: "必须启用"
        format: "int64"
      
      is_add:
        type: "integer"
        description: "是否加减项"
        format: "int64"
      
      f_enabled_mark:
        type: "integer"
        description: "有效标志0否1是"
        format: "int64"
      
      f_create_user_id:
        type: "string"
        description: "创建人ID"
        format: "string"
      
      f_modify_user_name:
        type: "string"
        description: "编辑人"
        format: "string"
      
      f_modify_user_id:
        type: "string"
        description: "编辑人ID"
        format: "string"
      
      f_delete_mark:
        type: "integer"
        description: "删除标记0否1是"
        format: "int64"
      
      f_sort_code:
        type: "integer"
        description: "排序码"
        format: "int64"
      
      f_modify_date:
        type: "string"
        description: "编辑日期"
        format: "date-time"
      
      f_create_user_name:
        type: "string"
        description: "创建人"
        format: "string"
      
      f_create_date:
        type: "string"
        description: "创建时间"
        format: "date-time"
      
      f_description:
        type: "string"
        description: "备注"
        format: "string"
  BasepayrollPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/Basepayroll"
...
  Personpayroll:
    type: "object"
    properties:
    
      id:
        type: "string"
        description: "ID"
        format: "string"
      
      tenant_id:
        type: "string"
        description: "租户编号"
        format: "string"
      
      person_id:
        type: "string"
        description: "雇员编号"
        format: "string"
      
      monthly_pay_amount:
        type: "number"
        description: "月薪"
        format: "float"
      
      valid_date:
        type: "string"
        description: "生效日期"
        format: "date-time"
      
      invalid_date:
        type: "string"
        description: "失效日期"
        format: "date-time"
      
      basic_sallary_amount:
        type: "number"
        description: "基本工资"
        format: "float"
      
      kpiamout:
        type: "number"
        description: "绩效工资"
        format: "float"
      
      traffic_fee:
        type: "number"
        description: "交通补贴"
        format: "float"
      
      food_fee:
        type: "number"
        description: "餐费补贴"
        format: "float"
      
      att_award_fee:
        type: "number"
        description: "全勤奖"
        format: "float"
      
      att_fee:
        type: "number"
        description: "考勤奖励"
        format: "float"
      
      professiona_fee:
        type: "number"
        description: "职称补贴"
        format: "float"
      
      job_allowance:
        type: "number"
        description: "岗位津贴"
        format: "float"
      
      assign_fee:
        type: "number"
        description: "外派项目补贴"
        format: "float"
      
      mobile_fee:
        type: "number"
        description: "通讯补贴"
        format: "float"
      
      she_bao_fee:
        type: "number"
        description: "社保标准"
        format: "float"
      
      house_fee:
        type: "number"
        description: "住房公积金"
        format: "float"
      
      cust_add_fee2:
        type: "number"
        description: "自定义加项2"
        format: "float"
      
      cust_add_fee1:
        type: "number"
        description: "自定义加项1"
        format: "float"
      
      cust_sub_fee2:
        type: "number"
        description: "自定义减项1"
        format: "float"
      
      cust_sub_fee1:
        type: "number"
        description: "自定义减项2"
        format: "float"
      
      is_last:
        type: "integer"
        description: "是否最新"
        format: "int64"
      
      spec_job_allowance:
        type: "number"
        description: "特殊岗位津贴"
        format: "float"
      
      she_bao_account_no:
        type: "string"
        description: "社保账号"
        format: "string"
      
      house_account_no:
        type: "string"
        description: "住房公积金账号"
        format: "string"
      
      f_enabled_mark:
        type: "integer"
        description: "有效标志0否1是"
        format: "int64"
      
      f_create_user_id:
        type: "string"
        description: "创建人ID"
        format: "string"
      
      f_modify_user_name:
        type: "string"
        description: "编辑人"
        format: "string"
      
      f_modify_user_id:
        type: "string"
        description: "编辑人ID"
        format: "string"
      
      f_delete_mark:
        type: "integer"
        description: "删除标记0否1是"
        format: "int64"
      
      f_sort_code:
        type: "integer"
        description: "排序码"
        format: "int64"
      
      f_modify_date:
        type: "string"
        description: "编辑日期"
        format: "date-time"
      
      f_create_user_name:
        type: "string"
        description: "创建人"
        format: "string"
      
      f_create_date:
        type: "string"
        description: "创建时间"
        format: "date-time"
      
      f_description:
        type: "string"
        description: "备注"
        format: "string"
  PersonpayrollPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/Personpayroll"
  Tenant:
    type: "object"
    properties:
    
      tenant_id:
        type: "string"
        description: "租户编号"
        format: "string"
      
      full_name:
        type: "string"
        description: "全称"
        format: "string"
      
      simple_name:
        type: "string"
        description: "简称"
        format: "string"
      
      company_id:
        type: "string"
        description: "公司系统编号"
        format: "string"
      
      contact_name:
        type: "string"
        description: "联系人"
        format: "string"
      
      contact_title:
        type: "string"
        description: "称谓"
        format: "string"
      
      contact_email:
        type: "string"
        description: "联系人邮件"
        format: "string"
      
      contact_phone:
        type: "string"
        description: "联系电话"
        format: "string"
      
      contact_mobile:
        type: "string"
        description: "联系人手机"
        format: "string"
      
      payroll_day:
        type: "integer"
        description: "薪资周期"
        format: "int64"
      
      hrsource:
        type: "string"
        description: "薪资周期"
        format: "string"
      
      att_source:
        type: "string"
        description: "人事来源"
        format: "string"
      
      buy_she_bao:
        type: "string"
        description: "社保设置"
        format: "string"
      
      ding_id:
        type: "string"
        description: "钉钉ID"
        format: "string"
      
      corpid:
        type: "string"
        description: "Corpid"
        format: "string"
      
      corpsecret:
        type: "string"
        description: "Corpsecret"
        format: "string"
      
      state:
        type: "integer"
        description: "租用状态"
        format: "int64"
      
      she_bao_rate:
        type: "number"
        description: "社保费率"
        format: "float"
      
      house_rate:
        type: "number"
        description: "公积金费率"
        format: "float"
      
      tenant_end_time:
        type: "string"
        description: "结束时间"
        format: "date-time"
      
      tenant_start_time:
        type: "string"
        description: "开始时间"
        format: "date-time"
      
      f_enabled_mark:
        type: "integer"
        description: "有效标志0否1是"
        format: "int64"
      
      f_create_user_id:
        type: "string"
        description: "创建人ID"
        format: "string"
      
      f_modify_user_name:
        type: "string"
        description: "编辑人"
        format: "string"
      
      f_modify_user_id:
        type: "string"
        description: "编辑人ID"
        format: "string"
      
      f_delete_mark:
        type: "integer"
        description: "删除标记0否1是"
        format: "int64"
      
      f_sort_code:
        type: "integer"
        description: "排序码"
        format: "int64"
      
      f_modify_date:
        type: "string"
        description: "编辑日期"
        format: "date-time"
      
      f_create_user_name:
        type: "string"
        description: "创建人"
        format: "string"
      
      f_create_date:
        type: "string"
        description: "创建时间"
        format: "date-time"
      
      f_description:
        type: "string"
        description: "备注"
        format: "string"
  TenantPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/Tenant"
  Tenantadjustitem:
    type: "object"
    properties:
    
      id:
        type: "string"
        description: "编号"
        format: "string"
      
      tenant_id:
        type: "string"
        description: "商户编号"
        format: "string"
      
      person_id:
        type: "string"
        description: "员工Id"
        format: "string"
      
      person_name:
        type: "string"
        description: "员工姓名"
        format: "string"
      
      dept_name:
        type: "string"
        description: "部门名称"
        format: "string"
      
      month_id:
        type: "string"
        description: "月份编号"
        format: "string"
      
      amount:
        type: "number"
        description: "调整金额"
        format: "float"
      
      adjest_remark:
        type: "string"
        description: "原因"
        format: "string"
      
      f_enabled_mark:
        type: "integer"
        description: "有效标志0否1是"
        format: "int64"
      
      f_create_user_id:
        type: "string"
        description: "创建人ID"
        format: "string"
      
      f_modify_user_name:
        type: "string"
        description: "编辑人"
        format: "string"
      
      f_modify_user_id:
        type: "string"
        description: "编辑人ID"
        format: "string"
      
      f_delete_mark:
        type: "integer"
        description: "删除标记0否1是"
        format: "int64"
      
      f_sort_code:
        type: "integer"
        description: "排序码"
        format: "int64"
      
      f_modify_date:
        type: "string"
        description: "编辑日期"
        format: "date-time"
      
      f_create_user_name:
        type: "string"
        description: "创建人"
        format: "string"
      
      f_create_date:
        type: "string"
        description: "创建时间"
        format: "date-time"
      
      f_description:
        type: "string"
        description: "备注"
        format: "string"
  TenantadjustitemPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/Tenantadjustitem"
...
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      msg:
        type: "string"
...